feat(ListAssets): logic for loading data has been moved out of the component
