+++ apps/web/src/app/[lang]/dho/[id]/treasury/page.tsx
@@ -3,7 +3,7 @@ import { Locale } from "@hypha-platform/i18n";
 import { Tabs, TabsList, TabsTrigger, TabsContent } from "@hypha-platform/ui/server";
 import Link from "next/link";
 import { ListAssets, ListPayouts, ListRequests } from "@hypha-platform/epics";
-import { listRequestsData, listPayoutsData } from "@hypha-platform/ui-utils";
+import { listRequestsData, listPayoutsData, listAssetsData } from "@hypha-platform/ui-utils";
 import { useState } from "react";
 
 type PageProps = {
@@ -13,6 +13,7 @@ type PageProps = {
 export default function TreasuryPage({ params: { lang, id } }: PageProps) {
   const [requests, setRequests] = useState(listRequestsData.requests);
   const [payouts, setPayouts] = useState(listPayoutsData.payouts);
+  const [assets, setAssets] = useState(listAssetsData.assets);
 
   const loadMoreRequests = () => {
     const newRequests = listRequestsData.newRequests;
@@ -22,6 +23,11 @@ export default function TreasuryPage({ params: { lang, id } }: PageProps) {
   const loadMorePayouts = () => {
     const newPayouts = listPayoutsData.newPayouts;
     setPayouts(prevRequests => [...prevRequests, ...newPayouts]);
+  }
+
+  const loadMoreAssets = () => {
+    const newAssets = listAssetsData.newAssets;
+    setAssets(prevRequests => [...prevRequests, ...newAssets]);
   }
 
   return (
@@ -46,7 +52,11 @@ export default function TreasuryPage({ params: { lang, id } }: PageProps) {
         </TabsList>
         <TabsContent value="treasury">
           <div className='flex flex-col justify-between items-center mt-4'>
-            <ListAssets/>
+            <ListAssets
+              assets={assets}
+              balance={listAssetsData.balance}
+              onLoadMore={loadMoreAssets}
+            />
             <ListRequests
               requests={requests}
               totalValue={listRequestsData.totalValue}
